requirements.txt
customtkinter~=5.2.2
pillow~=10.3.0
numpy~=2.0.0rc2